ATOM/huggingface/transformers/text2text-generation/pegasus/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
optimum-rbln
sentencepiece>=0.2.0
